onnx_app/requirements.txt
Flask
numpy
pillow
torchvision
onnx
onnxruntime